Rename src/lib/vm/vm.rs to src/lib/vm/core.rs.
It's a bit confusing to have a package and a module with the same name, so this
commit renames the module. Admittedly, `core.rs` is not very descriptive, though
it is probably accurate.
