+++ src/lib/vm/core.rs
