+++ src/lib/vm/objects/class.rs
@@ -16,9 +16,9 @@ use abgc_derive::GcLayout;
 use crate::{
     compiler::{cobjects, instrs::Instr},
     vm::{
+        core::{VMError, VM},
         objects::{Method, MethodBody, Obj, ObjType, StaticObjType, String_},
         val::{NotUnboxable, Val, ValResult},
-        vm::{VMError, VM},
     },
 };
